Enable funds to be funnelled for liquidity direction
