+++ test/voting/FundManager.ts
